Resize time data type for time structure
- hour, minute, second

+++ include/apps/watchfaces/OswAppWatchfaceDigital.h
@@ -21,7 +21,7 @@ class OswAppWatchfaceDigital: public OswAppV2 {
     static void refreshDateFormatCache();
     static void drawSteps();
     static void digitalWatch(short timeZone, uint8_t fontSize, uint8_t dateCoordY, uint8_t timeCoordY);
-    static void timeOutput(uint32_t hour, uint32_t minute, uint32_t second, bool showSecond = true);
+    static void timeOutput(uint8_t hour, uint8_t minute, uint8_t second, bool showSecond = true);
     static void dateOutput(uint16_t yearInt, uint8_t monthInt, uint8_t dayInt);
     static void displayWeekDay3(const char* weekday);
